+++ analysis/src/analyzer/sub_analyzer.rs
@@ -1,0 +1,5 @@
+use crate::analyzer::analyzer::Analyzer;
+
+pub trait SubAnalyzer: Sync + Send {
+  fn process_audio_chunk(&mut self, base_analyzer: &mut Analyzer, chunk: &[f32], force: bool);
+}
