+++ analysis/src/legacy/legacy_fft.rs
@@ -2,6 +2,9 @@ use log::{debug, info};
 
 use rubato::Resampler;
 use rubato::SincFixedIn;
+use rubato::SincInterpolationParameters;
+use rubato::SincInterpolationType;
+use rubato::WindowFunction;
 use rustfft::{num_complex::Complex, FftPlanner};
 use symphonia::core::audio::AudioBufferRef;
 use symphonia::core::audio::Signal;
@@ -10,11 +13,14 @@ use symphonia::core::conv::IntoSample;
 use symphonia::core::errors::Error;
 use tokio_util::sync::CancellationToken;
 
-use crate::features::energy;
-use crate::features::rms;
-use crate::features::zcr;
+use crate::utils::analyzer_utils::{build_hanning_window, AudioDescription};
+use crate::shared_utils::computing_device_type::ComputingDevice;
+use crate::utils::features::energy;
+use crate::utils::features::rms;
+use crate::utils::features::zcr;
 
-use crate::fft_utils::*;
+use crate::shared_utils::analyzer_shared_utils::*;
+use crate::wgpu_fft::wgpu_radix4;
 
 // Define the macro at the beginning of the file, after the imports
 macro_rules! process_window {
@@ -56,6 +62,14 @@ macro_rules! process_window {
         $sample_buffer.drain(..($window_size - $overlap_size));
     };
 }
+
+pub const RESAMPLER_PARAMETER: rubato::SincInterpolationParameters = SincInterpolationParameters {
+    sinc_len: 256,
+    f_cutoff: 0.95,
+    interpolation: SincInterpolationType::Linear,
+    oversampling_factor: 256,
+    window: WindowFunction::BlackmanHarris2,
+  };
 
 pub fn fft(
     file_path: &str,
