+++ analysis/src/shared_utils/computing_device_type.rs
