+++ cli/examples/build_database.rs
@@ -2,7 +2,7 @@ use log::{error, info};
 use std::path::PathBuf;
 use tracing_subscriber::filter::EnvFilter;
 
-use analysis::computing_device::ComputingDevice;
+use analysis::computing_device_type::ComputingDevice;
 use database::actions::analysis::{
     analysis_audio_library, empty_progress_callback as empty_analysis_progress_callback,
 };
